Renaming of tests related to Model component
