+++ src/test/java/seedu/task/model/task/NameTest.java
@@ -1,4 +1,4 @@
-package seedu.address.model.person;
+package seedu.task.model.task;
 
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
